clean up

+++ src/MPI.hpp
@@ -52,14 +52,14 @@ namespace Fractalium
         uint16_t width;
         uint16_t height;
 
-        uint64_t iterations{};
+        int iterations{};
 
         Double step_coord;
 
         Fractal fractal;
 
         MPIStruct(uint16_t start_x, uint16_t end_x, uint16_t start_y, uint16_t end_y, Double offset_x, Double offset_y,
-                  uint16_t width, uint16_t height, uint64_t iterations, Double step_coord, Fractal fractal) :
+                  uint16_t width, uint16_t height, int iterations, Double step_coord, Fractal fractal) :
                 start_x(start_x), end_x(end_x), start_y(start_y), end_y(end_y), offset_x(offset_x),
                 offset_y(offset_y), width(width), height(height), iterations(iterations), step_coord(step_coord),
                 fractal(fractal)
